+++ Lofas.SportsTracks.Hitta_SE_MapProvider/Common/CommonEnums.cs
@@ -11,7 +11,8 @@
     {
         Map,
         Aerial,
-        Nautical
+        Nautical,
+        Terrain
     }
         
 }